+++ quill/quill/include/quill/detail/LogManager.h
@@ -11,6 +11,9 @@
 #include "quill/detail/LoggerCollection.h"
 #include "quill/detail/ThreadContextCollection.h"
 
+#include "quill/detail/misc/Spinlock.h"
+#include "quill/detail/record/CommandRecord.h"
+
 namespace quill
 {
 namespace detail
@@ -69,17 +72,43 @@ public:
    * The backend thread will flush all loggers and all handlers up to the point (timestamp) that
    * this function was called.
    */
-  void flush();
+  void flush()
+  {
+    if (!_backend_worker.is_running())
+    {
+      // Backend worker needs to be running, otherwise we are stuck for ever waiting
+      return;
+    }
+
+    // Create an atomic variable
+    std::atomic<bool> backend_thread_flushed{false};
+
+    // notify will be invoked done the backend thread when this message is processed
+    auto notify_callback = [&backend_thread_flushed]() {
+      // When the backend thread is done flushing it will set the flag to true
+      backend_thread_flushed.store(true);
+    };
+
+    using log_record_t = detail::CommandRecord;
+    _thread_context_collection.local_thread_context()->spsc_queue().emplace<log_record_t>(notify_callback);
+
+    // The caller thread keeps checking the flag until the backend thread flushes
+    do
+    {
+      // wait
+      std::this_thread::sleep_for(std::chrono::nanoseconds{100});
+    } while (!backend_thread_flushed.load());
+  }
 
   /**
    * Starts the backend worker thread
    */
-  QUILL_ATTRIBUTE_COLD void start_backend_worker();
+  QUILL_ATTRIBUTE_COLD void inline start_backend_worker() { _backend_worker.run(); }
 
   /**
    * Stops the backend worker thread
    */
-  QUILL_ATTRIBUTE_COLD void stop_backend_worker();
+  QUILL_ATTRIBUTE_COLD void stop_backend_worker() { _backend_worker.stop(); }
 
 private:
   Config _config;