+++ quill/quill/include/quill/detail/TimestampFormatter.h
@@ -1,3 +1,8 @@
+/**
+ * Copyright(c) 2020-present, Odysseas Georgoudis & quill contributors.
+ * Distributed under the MIT License (http://opensource.org/licenses/MIT)
+ */
+
 #pragma once
 
 #include "quill/Fmt.h"