+++ quill/quill/include/quill/detail/misc/Common.h
@@ -11,9 +11,12 @@
 /**
  * Common type definitions etc
  */
+#if !defined(QUILL_RDTSC_CLOCK)
+  #define QUILL_RDTSC_CLOCK 1
+#endif
 
 #if !defined(QUILL_RDTSC_RESYNC_INTERVAL)
-  #define QUILL_RDTSC_RESYNC_INTERVAL 1
+  #define QUILL_RDTSC_RESYNC_INTERVAL 700
 #endif
 
 #if !defined(QUILL_ACTIVE_LOG_LEVEL)