+++ quill/quill/src/detail/misc/RdtscClock.cpp
@@ -1,7 +1,6 @@
 #include "quill/detail/misc/RdtscClock.h"
 
 #include "quill/detail/misc/Macros.h"
-#include "quill/detail/misc/Os.h"
 #include <algorithm>
 #include <array>
 #include <chrono>